lint

+++ packages/protocol/contracts/bridge/libs/LibBridgeProcess.sol
@@ -70,19 +70,15 @@ library LibBridgeProcess {
             status = IBridge.MessageStatus.DONE;
             refundAmount = message.callValue;
         } else {
-            uint gasLimit = msg.sender == message.owner ?
-                gasleft() :
-                message.gasLimit;
-            bool success = state.invokeMessageCall(
-                message,
-                mhash,
-                gasLimit
-            );
+            uint256 gasLimit = msg.sender == message.owner
+                ? gasleft()
+                : message.gasLimit;
+            bool success = state.invokeMessageCall(message, mhash, gasLimit);
 
             status = success
                 ? IBridge.MessageStatus.DONE
                 : IBridge.MessageStatus.RETRIABLE;
-        } 
+        }
 
         state.updateMessageStatus(mhash, status);
